# specs/fix-github-integration-and-update-constitution/contracts/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Physical AI & Humanoid Robotics Book Platform API",
    "description": "Backend API for the Physical AI & Humanoid Robotics book platform with free RAG chatbot, content management, and optional user features",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/user/notes": {
      "get": {
        "summary": "Get user notes",
        "tags": ["User"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "User notes",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object",
                    "properties": {
                      "id": {
                        "type": "integer"
                      },
                      "title": {
                        "type": "string"
                      },
                      "content": {
                        "type": "string"
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create user note",
        "tags": ["User"],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["title", "content", "content_id", "content_type"],
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "content": {
                    "type": "string"
                  },
                  "content_id": {
                    "type": "integer"
                  },
                  "content_type": {
                    "type": "string",
                    "enum": ["lesson", "lesson_section"]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Note created"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}